+++ uui-editor/src/components.tsx
@@ -3,6 +3,7 @@ import { MARK_BOLD, MARK_CODE, MARK_ITALIC, MARK_SUPERSCRIPT, MARK_UNDERLINE } f
 import { PlatePluginComponent } from '@udecode/plate-common';
 import { ELEMENT_H1, ELEMENT_H2, ELEMENT_H3, ELEMENT_H4, ELEMENT_H5, ELEMENT_H6 } from '@udecode/plate-heading';
 import { PARAGRAPH_KEY, ParagraphElement } from './plugins/paragraphPlugin';
+import { HeaderElement } from './plugins/headerPlugin/HeaderElement';
 
 export type DefaultPluginKey =
     | typeof ELEMENT_H1
@@ -24,12 +25,12 @@ export const createPlateUI = <T extends string = string>(
     >,
 ) => {
     const components: { [key: string]: PlatePluginComponent } = {
-        [ELEMENT_H1]: (props) => <h1 { ...props.attributes }>{ props.children }</h1>,
-        [ELEMENT_H2]: (props) => <h2 { ...props.attributes }>{ props.children }</h2>,
-        [ELEMENT_H3]: (props) => <h3 { ...props.attributes }>{ props.children }</h3>,
-        [ELEMENT_H4]: (props) => <h4 { ...props.attributes }>{ props.children }</h4>,
-        [ELEMENT_H5]: (props) => <h5 { ...props.attributes }>{ props.children }</h5>,
-        [ELEMENT_H6]: (props) => <h6 { ...props.attributes }>{ props.children }</h6>,
+        [ELEMENT_H1]: (props) => <HeaderElement variant="h1" { ...props } />,
+        [ELEMENT_H2]: (props) => <HeaderElement variant="h2" { ...props } />,
+        [ELEMENT_H3]: (props) => <HeaderElement variant="h3" { ...props } />,
+        [ELEMENT_H4]: (props) => <HeaderElement variant="h4" { ...props } />,
+        [ELEMENT_H5]: (props) => <HeaderElement variant="h5" { ...props } />,
+        [ELEMENT_H6]: (props) => <HeaderElement variant="h6" { ...props } />,
         [PARAGRAPH_KEY]: ParagraphElement,
         [MARK_BOLD]: (props) => <strong { ...props.attributes }>{ props.children }</strong>,
         [MARK_CODE]: (props) => <code { ...props.attributes }>{ props.children }</code>,