+++ changelog.md
@@ -2,11 +2,19 @@
 
 **What's New**
 * [RTE]:
-  * `paragraphPlugin`, `linkPlugin`:
+  * `paragraphPlugin`, `linkPlugin`, `headerPlugin`:
     * added full extendability support and types  
       Created with `createPluginFactory`, see [Plate docs / createPluginFactory](https://github.com/udecode/plate/blob/23d841bf57fef940ae8f4d20ee383474dbc6cdd6/apps/www/content/docs/api/core.mdx#createpluginfactory).
     * added `slate-` class name to plugin element  
       Now corresponds to [Plate docs / Styling with CSS](https://github.com/udecode/plate/blob/23d841bf57fef940ae8f4d20ee383474dbc6cdd6/apps/www/content/docs/plugin-components.mdx#styling-with-css).
+  * [Breaking change] `headerPlugin`: plugin options moved
+    ```js
+    // before
+    headerPlugin({ headers: [...] })
+    
+    // after
+    headerPlugin({ options: { headers: [...] } })
+    ```
 
 **What's Fixed**